add startup provisioning and scripts fix minor bug in shutdown

+++ server/dataman.c
@@ -190,7 +190,8 @@ void shutdown_handler(int sig)
 	
 void useage(char *name)
 {
-	fprintf(stderr, "%s: useage: %s [-D|m size|n num|q|s|t]\n"
+	fprintf(stderr, "%s: useage: %s [-D|f|m size|n num|q|s|t]\n"
+					"    -f  stay in foreground for a service manager\n"
 					"    -D  start with debugging turned on\n"
 					"            if dataman is already running will toggle the\n"
 					"            state of debugging\n"
@@ -211,6 +212,7 @@ int main(int argc, char *argv[])
 
 	int i, j;						/* loop counters */
 	int ssw, tsw;					/* control switches */
+	int foreground = 0;
 	int qsw;						/* query switch */
 	int file_opened;				/* another switch */
 
@@ -240,6 +242,9 @@ int main(int argc, char *argv[])
 			useage(argv[0]);
 		for (j = 1; j < strlen(argv[i]); j++) {
 			switch(argv[i][j]) {
+				case 'f':
+					foreground = 1;
+					break;
 				case 'D':
 					if (dbgsw ||qsw || ssw || tsw)
 						useage(argv[0]);
@@ -299,6 +304,17 @@ int main(int argc, char *argv[])
 /*
  * find out what of the other routines are running
  */
+	/* Service managers already own the process lifecycle. Acquire our PID
+	 * lock directly, without daemonizing or killing independently run children. */
+	if (foreground) {
+		if (qsw || ssw || tsw)
+			useage(argv[0]);
+		if (verify_pid(basename(argv[0])) < 0) {
+			perror("dataman: cannot own supervisor PID file");
+			return EXIT_FAILURE;
+		}
+		goto start_services;
+	}
 	file_opened = 0;
 	if ((fp = popen("ls /tmp/.dataman*.pid 2>/dev/null", "r")) == NULL) {
 		fprintf(stderr, "Can't popen for current status: ");
@@ -425,6 +441,7 @@ int main(int argc, char *argv[])
 		exit(errno);
 
 	setsid();				/* become the session leader */
+start_services:
 	if (chdir("/tmp") < 0) {
 		fprintf(stderr, "%s: can't change to /tmp: ", argv[0]);
 		perror("");
@@ -435,7 +452,8 @@ int main(int argc, char *argv[])
  * at this point verify_pid should always return true, because we have
  * already determined that we aren't running.
  */
-	verify_pid(basename(argv[0]));
+	if (!foreground && verify_pid(basename(argv[0])) < 0)
+		return EXIT_FAILURE;
 /*
  * ok, now we are a daemon, set up the signal catchers for the child
  * processes
@@ -526,14 +544,16 @@ int main(int argc, char *argv[])
  * set up a logfile, close stdio channels and let the user know
  * that we are in good shape!
  */
-	fclose(stdin);
-	if ((fp = freopen("/tmp/dataman.log", "w+", stderr)) == NULL) {
-		kill(con_pid, SIGKILL);
-		kill(srv_pid, SIGKILL);
-		err_sys("%s: Can't redirect stderr: ", argv[0]);
+	if (!foreground) {
+		fclose(stdin);
+		if ((fp = freopen("/tmp/dataman.log", "w+", stderr)) == NULL) {
+			kill(con_pid, SIGKILL);
+			kill(srv_pid, SIGKILL);
+			err_sys("%s: Can't redirect stderr: ", argv[0]);
+		}
+		fprintf(stdout, "\n%s: running...\n", argv[0]);
+		fclose(stdout);
 	}
-	fprintf(stdout, "\n%s: running...\n", argv[0]);
-	fclose(stdout);
 /*
  * ok, now do nothing for ever
  */
